+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+
+
+## [v1.4.0] - 2025-04-20
 ### Added
 - Added `--headless/--no-headless` flag, which is mostly used for debugging.